GMM and mathsutilities refactored
